configure oauth client id per athenz service
