+++ servers/zms/schema/updates/update-20251119.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `zms_server`.`service` ADD COLUMN `client_id` VARCHAR(256) NOT NULL DEFAULT '';
