+++ servers/zms/schema/zms_server.sql
@@ -1,5 +1,5 @@
 -- MySQL Script generated by MySQL Workbench
--- Fri Aug  8 14:47:06 2025
+-- Wed Nov 19 19:20:35 2025
 -- Model: New Model    Version: 1.0
 -- MySQL Workbench Forward Engineering
 
@@ -193,6 +193,7 @@ CREATE TABLE IF NOT EXISTS `zms_server`.`service` (
   `x509_cert_signer_keyid` VARCHAR(64) NOT NULL DEFAULT '',
   `ssh_cert_signer_keyid` VARCHAR(64) NOT NULL DEFAULT '',
   `creds` VARCHAR(256) NOT NULL DEFAULT '',
+  `client_id` VARCHAR(256) NOT NULL DEFAULT '',
   PRIMARY KEY (`service_id`),
   INDEX `fk_service_domain_idx` (`domain_id` ASC),
   UNIQUE INDEX `uq_domain_service` (`name` ASC, `domain_id` ASC),
